feat: config .net in 9.0, connection mysql, migrations and repositories

+++ src/Itau.CompraProgramada.Domain/Entities/Entity.cs
@@ -4,7 +4,7 @@ namespace Itau.CompraProgramada.Domain.Entities
 {
     public abstract class Entity
     {
-        public int Id { get; protected set; }
+        public long Id { get; protected set; }
         public DateTime DataCriacao { get; protected set; } = DateTime.UtcNow;
     }
 }
